realtime_dynamic_plots_from_SQL_database/requirements_dev.txt
numpy==1.24.4
pandas==1.4.4
pyarrow==18.1.0
psycopg2-binary==2.9.9
requests==2.32.3